requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.8
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
idna==2.6
Markdown==2.6.11
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
requests==2.18.4
urllib3==1.22
